libfreerdp-rfx:
removing rfx_private.h and rfx_constants.h, moving definitions to rfx.h
