+++ libfreerdp-rfx/rfx.c
@@ -20,11 +20,10 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <freerdp/rfx.h>
 #include <freerdp/types/base.h>
 #include <freerdp/utils/stream.h>
 
-#include "rfx_constants.h"
-#include "rfx_private.h"
 #include "rfx_decode.h"
 
 RFX_CONTEXT *
